SCD4X/scd41-co2-monitor-local/local-sensor-api/requirements.txt
flask==3.0.0
psycopg[binary]==3.2.4
python-dotenv==1.0.0
